# apps/api/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/v1/auth/sign-in": {
      "post": {
        "operationId": "AuthController_signIn",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignInDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionCredentialsDto"
                }
              }
            }
          },
          "401": {
            "description": "Generic response for all invalid account credentials or state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "429": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "summary": "Sign in with an existing KCCA account",
        "tags": [
          "authentication"
        ]
      }
    },
    "/v1/auth/client-code/request": {
      "post": {
        "description": "Returns REGISTRATION_REQUIRED only for genuine phone absence. Existing ineligible accounts receive a generic error. Valid initial requests consume protected phone/IP throttles before lookup.",
        "operationId": "AuthController_requestClientCode",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientCodeRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/PhoneChallengeDto"
                    },
                    {
                      "$ref": "#/components/schemas/RegistrationRequiredDto"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or OTP cooldown; retryAt is supplied for cooldown.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "Generic unavailable response without account details.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "429": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "summary": "Request a Client sign-in code or registration",
        "tags": [
          "authentication"
        ]
      }
    },
    "/v1/auth/client-code/verify": {
      "post": {
        "operationId": "AuthController_verifyClientCode",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyPhoneCodeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionCredentialsDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "summary": "Sign in a Client with a phone verification code",
        "tags": [
          "authentication"
        ]
      }
    },
    "/v1/auth/client-code/resend": {
      "post": {
        "operationId": "AuthController_resendClientCode",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResendPhoneCodeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhoneChallengeDto"
                }
              }
            }
          },
          "400": {
            "description": "OTP cooldown with retryAt.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "429": {
            "description": "Hourly issuance limit with retryAt.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "summary": "Replace and resend a Client sign-in code",
        "tags": [
          "authentication"
        ]
      }
    },
    "/v1/auth/provider-code/request": {
      "post": {
        "description": "Verified registered phone and login eligibility required. Unknown and ineligible accounts receive the same guidance; no account is created.",
        "operationId": "AuthController_requestProviderCode",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProviderCodeRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhoneChallengeDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "429": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "summary": "Request a Service Provider sign-in code",
        "tags": [
          "authentication"
        ]
      }
    },
    "/v1/auth/provider-code/verify": {
      "post": {
        "operationId": "AuthController_verifyProviderCode",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyPhoneCodeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionCredentialsDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "summary": "Sign in a Provider with a phone verification code",
        "tags": [
          "authentication"
        ]
      }
    },
    "/v1/auth/provider-code/resend": {
      "post": {
        "operationId": "AuthController_resendProviderCode",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResendPhoneCodeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhoneChallengeDto"
                }
              }
            }
          },
          "400": {
            "description": "OTP cooldown with retryAt.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "429": {
            "description": "Hourly issuance limit with retryAt.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "summary": "Replace and resend a Provider sign-in code",
        "tags": [
          "authentication"
        ]
      }
    },
    "/v1/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionCredentialsDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "summary": "Rotate a Weyonje refresh token",
        "tags": [
          "authentication"
        ]
      }
    },
    "/v1/auth/sign-out": {
      "post": {
        "operationId": "AuthController_signOut",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignOutDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Revoke the current Weyonje session",
        "tags": [
          "authentication"
        ]
      }
    },
...
  },
  "info": {
    "title": "Weyonje API",
    "description": "Native authentication, server-managed sessions, and authoritative eligibility for Weyonje clients.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "PhoneChallengeDto": {
        "type": "object",
        "properties": {
          "challengeId": {
            "type": "string",
            "format": "uuid"
          },
          "maskedPhone": {
            "type": "string",
            "example": "+256 •••••• 123"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "resendAvailableAt": {
            "type": "string",
            "format": "date-time"
          },
          "deliveryStatus": {
            "type": "string",
            "enum": [
              "SENT",
              "FAILED"
            ]
          }
        },
        "required": [
          "challengeId",
          "maskedPhone",
          "expiresAt",
          "resendAvailableAt",
          "deliveryStatus"
        ]
      },
      "RegistrationRequiredDto": {
        "type": "object",
        "properties": {
          "outcome": {
            "type": "string",
            "enum": [
              "REGISTRATION_REQUIRED"
            ]
          }
        },
        "required": [
          "outcome"
        ]
      },
      "SignInDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "account@example.invalid"
          },
          "password": {
            "type": "string",
            "format": "password",
            "writeOnly": true
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "SessionCredentialsDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "accessTokenExpiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "refreshTokenExpiresAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "accessToken",
          "refreshToken",
          "accessTokenExpiresAt",
          "refreshTokenExpiresAt"
        ]
      },
      "ApiErrorDto": {
        "type": "object",
        "properties": {
          "retryAt": {
            "type": "string",
            "format": "date-time"
          },
          "limitCategory": {
            "type": "string",
            "enum": [
              "OTP_HOURLY",
              "OTP_COOLDOWN",
              "CLIENT_REQUEST",
              "PROVIDER_REQUEST"
            ]
          },
          "code": {
            "type": "string",
            "example": "AUTH_ACCESS_DENIED"
          },
          "message": {
            "type": "string",
            "example": "This account cannot access Weyonje mobile services."
          },
          "requestId": {
            "type": "string",
            "example": "2e96d4c8-5f83-4de0-957e-a7f0cff86690"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "ClientCodeRequestDto": {
        "type": "object",
        "properties": {
          "phoneNumber": {
            "type": "string",
            "example": "+256 700 000000"
          }
        },
        "required": [
          "phoneNumber"
        ]
      },
      "VerifyPhoneCodeDto": {
        "type": "object",
        "properties": {
          "challengeId": {
            "type": "string",
            "format": "uuid"
          },
          "code": {
            "type": "string",
            "pattern": "^[0-9]{6}$",
            "writeOnly": true
          }
        },
        "required": [
          "challengeId",
          "code"
        ]
      },
      "ResendPhoneCodeDto": {
        "type": "object",
        "properties": {
          "challengeId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "challengeId"
        ]
      },
      "ProviderCodeRequestDto": {
        "type": "object",
        "properties": {
          "phoneNumber": {
            "type": "string",
            "example": "+256 700 000000"
          }
        },
        "required": [
          "phoneNumber"
        ]
      },
      "RefreshDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "writeOnly": true
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "SignOutDto": {
        "type": "object",
        "properties": {
          "signedOut": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "signedOut"
        ]
      },
...
    }
  }
}